rename subscribe_tx_throughput
