+++ massa-grpc/src/handler.rs
@@ -4,9 +4,8 @@
 use massa_proto::massa::api::v1::{
     self as grpc, GetBlocksBySlotRequest, GetBlocksBySlotResponse, GetNextBlockBestParentsRequest,
     GetNextBlockBestParentsResponse, GetTransactionsThroughputRequest,
-    GetTransactionsThroughputResponse, GetTransactionsThroughputStreamRequest,
-    NewBlocksHeadersStreamRequest, NewBlocksStreamRequest, NewFilledBlocksStreamRequest,
-    NewOperationsStreamRequest,
+    GetTransactionsThroughputResponse, NewBlocksHeadersStreamRequest, NewBlocksStreamRequest,
+    NewFilledBlocksStreamRequest, NewOperationsStreamRequest, TransactionsThroughputStreamRequest,
 };
 
 use crate::api::{
@@ -18,9 +17,7 @@ use crate::stream::new_blocks::{new_blocks, NewBlocksStream};
 use crate::stream::new_blocks_headers::{new_blocks_headers, NewBlocksHeadersStream};
 use crate::stream::new_filled_blocks::{new_filled_blocks, NewFilledBlocksStream};
 use crate::stream::new_operations::{new_operations, NewOperationsStream};
-use crate::stream::subscribe_tx_throughput::{
-    subscribe_transactions_throughput, SubscribeTransactionsThroughputStream,
-};
+use crate::stream::tx_throughput::{transactions_throughput, TransactionsThroughputStream};
 use crate::stream::{
     send_blocks::{send_blocks, SendBlocksStream},
     send_endorsements::{send_endorsements, SendEndorsementsStream},
@@ -134,14 +131,14 @@ impl grpc::grpc_server::Grpc for MassaGrpcService {
         }
     }
 
-    type SubscribeTransactionsThroughputStream = SubscribeTransactionsThroughputStream;
+    type TransactionsThroughputStream = TransactionsThroughputStream;
 
-    /// Handler for subscribe on transactions throughput
-    async fn subscribe_transactions_throughput(
+    /// Handler for transactions throughput
+    async fn transactions_throughput(
         &self,
-        request: Request<Streaming<GetTransactionsThroughputStreamRequest>>,
-    ) -> Result<Response<Self::SubscribeTransactionsThroughputStream>, Status> {
-        match subscribe_transactions_throughput(self, request).await {
+        request: Request<Streaming<TransactionsThroughputStreamRequest>>,
+    ) -> Result<Response<Self::TransactionsThroughputStream>, Status> {
+        match transactions_throughput(self, request).await {
             Ok(res) => Ok(Response::new(res)),
             Err(e) => Err(e.into()),
         }
